+++ src/js/core/methods/AbstractMethod.js
@@ -1,16 +1,15 @@
 /* @flow */
-'use strict';
+import { crypto } from 'bitcoinjs-lib-zcash';
 
 import Device from '../../device/Device';
-
+import DataManager from '../../data/DataManager';
 import * as UI from '../../constants/ui';
 import * as DEVICE from '../../constants/device';
+import { load as loadStorage, save as saveStorage, PERMISSIONS_KEY } from '../../iframe/storage';
+
 import { UiMessage, DeviceMessage } from '../../message/builder';
 import type { Deferred, CoreMessage, UiPromiseResponse } from '../../types';
-
-import { load as loadStorage, save as saveStorage, PERMISSIONS_KEY } from '../../iframe/storage';
-import { crypto } from 'bitcoinjs-lib-zcash';
-import DataManager from '../../data/DataManager';
+import type { ButtonRequestData } from '../../types/ui-request';
 
 export interface MethodInterface {
     +responseID: number,
@@ -40,6 +39,7 @@ export default class AbstractMethod implements MethodInterface {
     allowDeviceMode: Array<string>; // used in device management (like ResetDevice allow !UI.INITIALIZED)
 
     +confirmation: () => Promise<boolean>;
+    +getButtonRequestData: (code: string) => ?ButtonRequestData;
 
     // // callbacks
     postMessage: (message: CoreMessage) => void;