+++ src/js/core/methods/CardanoGetAddress.js
@@ -1,5 +1,4 @@
 /* @flow */
-'use strict';
 
 import AbstractMethod from './AbstractMethod';
 import { validateParams } from './helpers/paramsValidator';
@@ -8,50 +7,43 @@ import { validatePath, fromHardened, getSerializedPath } from '../../utils/pathU
 import * as UI from '../../constants/ui';
 import { UiMessage } from '../../message/builder';
 
-import type { CardanoAddress } from '../../types/trezor';
 import type { CoreMessage, UiPromiseResponse } from '../../types';
-import type { CardanoAddress as CardanoAddressResponse } from '../../types/cardano';
+import type { CardanoAddress } from '../../types/cardano';
 
 type Batch = {
     path: Array<number>,
+    address: ?string,
     showOnTrezor: boolean,
 }
 
-type Params = {
-    bundle: Array<Batch>,
-    bundledResponse: boolean,
-}
+type Params = Array<Batch>;
 
 export default class CardanoGetAddress extends AbstractMethod {
-    params: Params;
     confirmed: boolean = false;
+    params: Params;
+    progress: number = 0;
 
     constructor(message: CoreMessage) {
         super(message);
 
         this.requiredPermissions = ['read'];
         this.requiredFirmware = ['0', '2.0.8'];
-        this.info = 'Export Cardano address';
 
-        const payload: Object = message.payload;
-        let bundledResponse: boolean = true;
-        // create a bundle with only one batch
-        if (!payload.hasOwnProperty('bundle')) {
-            payload.bundle = [ ...payload ];
-            bundledResponse = false;
-        }
+        // create a bundle with only one batch if bundle doesn't exists
+        const payload: Object = !message.payload.hasOwnProperty('bundle') ? { ...message.payload, bundle: [ ...message.payload ] } : message.payload;
 
         // validate bundle type
         validateParams(payload, [
             { name: 'bundle', type: 'array' },
+            { name: 'useEventListener', type: 'boolean' },
         ]);
 
         const bundle = [];
-
         payload.bundle.forEach(batch => {
             // validate incoming parameters for each batch
             validateParams(batch, [
                 { name: 'path', obligatory: true },
+                { name: 'address', type: 'string' },
                 { name: 'showOnTrezor', type: 'boolean' },
             ]);
 
@@ -63,14 +55,34 @@ export default class CardanoGetAddress extends AbstractMethod {
 
             bundle.push({
                 path,
+                address: batch.address,
                 showOnTrezor,
             });
         });
 
-        this.params = {
-            bundle,
-            bundledResponse,
-        };
+        const useEventListener = payload.useEventListener && payload.bundle.length === 1 && typeof payload.bundle[0].address === 'string' && payload.bundle[0].showOnTrezor;
+        this.confirmed = useEventListener;
+        this.useUi = !useEventListener;
+        this.params = bundle;
+
+        // set info
+        if (bundle.length === 1) {
+            this.info = `Export Cardano address for account #${ (fromHardened(this.params[0].path[2]) + 1) }`;
+        } else {
+            this.info = 'Export multiple Cardano addresses';
+        }
+    }
+
+    getButtonRequestData(code: string) {
+        if (code === 'ButtonRequest_Address') {
+            const data = {
+                type: 'address',
+                serializedPath: getSerializedPath(this.params[this.progress].path),
+                address: this.params[this.progress].address || 'not-set',
+            };
+            return data;
+        }
+        return null;
     }
 
     async confirmation(): Promise<boolean> {
@@ -80,13 +92,7 @@ export default class CardanoGetAddress extends AbstractMethod {
         // initialize user response promise
         const uiPromise = this.createUiPromise(UI.RECEIVE_CONFIRMATION, this.device);
 
-        let label: string;
-        if (this.params.bundle.length > 1) {
-            label = 'Export multiple Cardano addresses';
-        } else {
-            label = `Export Cardano address for account #${ (fromHardened(this.params.bundle[0].path[2]) + 1) }`;
-        }
-
+        const label: string = this.info;
         // request confirmation view
         this.postMessage(new UiMessage(UI.REQUEST_CONFIRMATION, {
             view: 'export-address',
@@ -101,11 +107,29 @@ export default class CardanoGetAddress extends AbstractMethod {
         return this.confirmed;
     }
 
-    async run(): Promise<CardanoAddressResponse | Array<CardanoAddressResponse>> {
-        const responses: Array<CardanoAddressResponse> = [];
-        for (let i = 0; i < this.params.bundle.length; i++) {
-            const batch: Batch = this.params.bundle[i];
-            const response: CardanoAddress = await this.device.getCommands().cardanoGetAddress(
+    async run(): Promise<CardanoAddress | Array<CardanoAddress>> {
+        const responses: Array<CardanoAddress> = [];
+        const bundledResponse = this.params.length > 1;
+
+        for (let i = 0; i < this.params.length; i++) {
+            const batch: Batch = this.params[i];
+            // silently get address and compare with requested address
+            // or display as default inside popup
+            if (batch.showOnTrezor) {
+                const silent = await this.device.getCommands().cardanoGetAddress(
+                    batch.path,
+                    false
+                );
+                if (typeof batch.address === 'string') {
+                    if (batch.address !== silent.address) {
+                        throw new Error('Addresses do not match');
+                    }
+                } else {
+                    batch.address = silent.address;
+                }
+            }
+
+            const response = await this.device.getCommands().cardanoGetAddress(
                 batch.path,
                 batch.showOnTrezor
             );
@@ -115,14 +139,16 @@ export default class CardanoGetAddress extends AbstractMethod {
                 address: response.address,
             });
 
-            if (this.params.bundledResponse) {
+            if (bundledResponse) {
                 // send progress
                 this.postMessage(new UiMessage(UI.BUNDLE_PROGRESS, {
                     progress: i,
                     response,
                 }));
             }
+
+            this.progress++;
         }
-        return this.params.bundledResponse ? responses : responses[0];
+        return bundledResponse ? responses : responses[0];
     }
 }