+++ src/js/data/DataManager.js
@@ -71,7 +71,7 @@ export default class DataManager {
             // check if origin is localhost or trusted
             const isLocalhost: boolean = typeof window !== 'undefined' ? window.location.hostname === 'localhost' : true;
             const whitelist: ?WhiteList = DataManager.isWhitelisted(this.settings.origin || '');
-            this.settings.trustedHost = isLocalhost || (!!whitelist && !this.settings.popup);
+            this.settings.trustedHost = (isLocalhost || !!whitelist) && !this.settings.popup;
             // ensure that popup will be used
             if (!this.settings.trustedHost) {
                 this.settings.popup = true;