+++ src/js/types/index.js
@@ -125,6 +125,7 @@ export type UiMessage = {
         device: Device,
         code?: string,
         browser?: any,
+        data?: any,
     },
 }